fix(CI migrations): It doesn't migrate data in production

+++ html/database/migrations/Production.php
@@ -9,16 +9,22 @@ trait Production
 {
     public function upInLocalOrProduction($name, $create, $update = null)
     {
-        Schema::create($name, $create);
+        if (!App::environment('production')) {
+            Schema::create($name, $create);
+        }
     }
 
     public function dropInLocalNoProduction($name)
     {
-        Schema::dropIfExists($name);
+        if (!App::environment('production')) {
+            Schema::dropIfExists($name);
+        }
     }
 
     public function downInLocalOrProduction($destroy, $update = null)
     {
-        $destroy();
+        if (!App::environment('production')) {
+            $destroy();
+        }
     }
 }
